requirements.txt
# AWM System Dependencies

# Core Python packages
asyncio-mqtt==0.16.1
asyncpg==0.29.0
redis==5.0.1
aiohttp==3.9.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Technical analysis
TA-Lib==0.4.28
pandas-ta==0.3.14b0

# Machine Learning / AI
scikit-learn==1.3.2
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Web framework
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2

# API clients
requests==2.31.0
websockets==12.0
kiteconnect==4.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
typer==0.9.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Notifications
python-telegram-bot==20.7
slack-sdk==3.26.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
